// src/ambrosia/ClientApp/store/recipeform/mutation-types.ts
export enum RecipeFormMutation {
    EditName = 'editName',
    EditDescription = 'editDescription',
    EditCookTime = 'editCookTime',
    EditPrepTime = 'editPrepTime',
    AddStep = 'addStep',
    EditStep = 'editStep',
    RemoveStep = 'removeStep',
    AddIngredient = 'addIngredient',
    EditIngredient = 'editIngredient',
    RemoveIngredient = 'removeIngredient',
}